Change Group Data Settings tab to Privacy & Access
Move all visibility and access related settings to there
