+++ src/css/global/_sass_resources.scss
@@ -97,7 +97,7 @@ $color-cod-gray-0: rgba(#161616, 0);
 $color-cod-gray-dark: #1C1C1C;
 $color-seashell: #F1F1F1;
 $color-catskill-white: #F9FAFC;
-$color-slate-gray-80: rgba(#67768A, .6);
+$color-slate-gray-80: rgba(#67768A, .7);
 $color-saltpan: #FAFCFB;
 $color-loblolly: #BFC5CD;
 $color-box-shadow: rgba(0, 0, 0, 0.15);
