--- src/routes/GroupSettings/DataSettingsTab/index.js
@@ -1,3 +1,0 @@
-import component from './DataSettingsTab'
-
-export default component
