+++ src/routes/GroupSettings/GroupSettingsTab/GroupSettingsTab.scss
@@ -40,314 +40,8 @@
   width: 30px;
 }
 
-.privacy-settings {
-  margin: 25px 0;
-}
-
-.privacyDetail {
-  color: $color-regent;
-  font-size: 14px;
-  margin: 0;
-}
-
-.privacy-option-container {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  height: 40px;
-  line-height: 1;
-  margin-bottom: 25px;
-}
-
-.privacy-option.disabled {
-  color: $color-regent-70;
-}
-
-label {
-  color: $color-regent-70;
-  // LEJ: Commenting-out as has global impact
-  // margin-left: 10px;
-  transition: color .25s ease;
-}
-
-.privacy-state {
-  border: 1px solid $color-regent;
-  border-radius: 10px;
-  color: $color-regent;
-  font-size: 10px;
-  font-style: bold;
-  transition: color .25s ease;
-  padding: 2px 6px;
-}
-
 .skillsSetting {
   padding: 0;
   position: relative;
-  padding-bottom: 25px;
   margin-left: -15px;
-}
-
-.toggleDescription {
-  display: inline;
-  font-size: 14px;
-  color: rgba(34, 64, 91, .5);
-  padding-left: 15px;
-}
-.on > .switchContainer .toggleDescription {
-  color: rgba(34, 64, 91, 1);
-}
-
-.privacySetting {
-  padding: 10px 0;
-
-  label {
-    display: flex;
-    align-items: center;
-
-    .settingIcon {
-      margin: 0 40px 0 20px;
-      width: 0px;
-    }
-
-    h4 {
-      font-size: 18px;
-      margin-bottom: 0;
-      transition: all .25s ease;
-    }
-
-    .privacy-option {
-      transition: all .25s ease;
-      color: rgba(125, 140, 157, 1.000);
-      font-size: 14px;
-      margin-left: 0;
-      flex-grow: 1;
-    }
-
-    .settingDescription {
-      line-height: 18px;
-    }
-
-  }
-}
-
-.prerequisiteWarning {
-  margin-top: -15px;
-
-  .prerequisiteIcon {
-    font-size: 16px;
-    position: relative;
-    top: 3px;
-    color: rgba(41, 64, 90, 1.00);
-    margin: 0 3px;
-  }
-
-  strong {
-    color: rgba(41, 64, 90, 1.00);
-  }
-
-  .warning {
-    color: rgba(237, 86, 83, 1.00);
-  }
-}
-
-.onOff {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  right: 0;
-
-  div {
-    font-size: 10px;
-    height: 20px;
-    border-radius: 20px;
-    border: 1px solid rgba(125, 140, 157, 1.000);
-    color: rgba(125, 140, 157, 1.000);
-    padding: 0 10px;
-    line-height: 20px;
-  }
-
-  .on {
-    display: none;
-    border: 1px solid rgba(0, 163, 227, 1.000);
-    color: rgba(0, 163, 227, 1.000);
-  }
-}
-
-.on > div > .onOff {
-  .off {
-    display: none;
-  }
-
-  .on {
-    display: block;
-  }
-}
-
-.privacySetting.on {
-  .privacy-state {
-    border: 1px solid $color-picton-blue;
-    color: $color-picton-blue;
-  }
-
-  label {
-    color: $color-regent !important;
-
-    .settingIcon {
-      color: rgba(0, 117, 255, 1.000);
-    }
-
-    h4, .privacy-option {
-      color: rgba(35, 64, 91, 1.000);
-    }
-  }
-}
-
-.saveChanges {
-  position: fixed;
-  bottom: 10px;
-  padding: 10px;
-  background: rgba(255,255,255,1);
-  box-shadow: 0px 4px 15px rgba(0, 0, 0, 0.15);
-  border-radius: 5px;
-  display: flex;
-  width: inherit;
-  max-width: 95%;
-  margin-top: $space-9x;
-  align-items: center;
-  justify-content: space-between;
-
-  span {
-    display: inline-block;
-    color: rgba(35, 64, 91, .5);
-    margin-right: 30px;
-    font-size: 14px;
-  }
-
-  span.settingChanged {
-    color: rgba(35, 64, 91, 1);
-  }
-
-  .save-button {
-    margin: 0 auto;
-  }
-}
-
-.switchContainer {
-  position: relative;
-  display: flex;
-  align-items: center;
-  padding-right: 40px;
-  margin-bottom: 20px;
-}
-
-.on .settingDisabled {
-  opacity: 0;
-}
-
-.groupQuestions {
-  position: relative;
-  margin-left: 13px;
-  padding-top: 10px;
-  transition: all .25s ease;
-
-  .questionList {
-    margin-top: 5px;
-    margin-left: 58px;
-
-    span {
-      transition: all .25s ease;
-      color: rgba(125, 140, 157, 1.000);
-      font-size: 14px;
-      line-height: 18px;
-      position: relative;
-      top: -2px;
-      display: block;
-      margin-bottom: 10px;
-    }
-  }
-
-  .question {
-    opacity: .5;
-    transition: all .25s ease;
-    margin-bottom: 8px;
-    background: #FFFFFF;
-    box-shadow: 0px 4px 15px rgba(0, 0, 0, 0.05);
-    border-radius: 5px;
-    position: relative;
-
-    .deleteInput {
-      position: absolute;
-      top: 8px;
-      right: 6px;
-      font-size: 18px;
-
-      .close {
-        color: rgba(255, 93, 99, 1.000);
-        font-size: 18px;
-        cursor: pointer;
-      }
-    }
-
-    .createInput {
-      position: absolute;
-      top: 5px;
-      font-size: 20px;
-      right: 10px;
-    }
-
-    input {
-      width: 100%;
-      border: none;
-      font-size: 14px;
-      padding: 5px 25px 5px 8px;
-      color: rgba(35, 64, 91, 1.000);
-      border-radius: 3px;
-
-      &::placecholder {
-        font-style: italic;
-      }
-    }
-  }
-}
-
-.groupQuestions.on {
-  .questionList span {
-    color: rgba(35, 64, 91, 1.000);
-  }
-
-  .question {
-    opacity: 1;
-  }
-}
-
-.member-visibility {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.groupSettings {
-  padding-bottom: 150px;
-}
-
-@media screen and (max-width: 425px) {
-  .save-button {
-    font-size: 12px;
-  }
-}
-
-@media screen and (max-width: 600px) {
-  .saveChanges {
-    bottom: 60px;
-
-    span {
-      font-size: 12px;
-    }
-
-    .save-button {
-      font-size: 12px;
-      min-width: 120px;
-      text-align: center;
-    }
-  }
 }
