+++ src/routes/GroupSettings/PrivacySettingsTab/index.js
@@ -1,0 +1,3 @@
+import component from './PrivacySettingsTab'
+
+export default component
